api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
requests==2.31.0
python-multipart==0.0.6
# Añadir urllib3 compatible
urllib3==1.26.18
# Añadir certificados SSL
certifi>=2023.7.22

